requirements.txt
requests==2.22.0
protobuf==3.13.0
google-cloud-firestore>=1.9.0